requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
google-generativeai
sqlalchemy
pgvector
psycopg[binary]
pypdf